# perf(ext/web): optimize atob/btoa