--- a/ext/web/05_base64.js
+++ b/ext/web/05_base64.js
@@ -9,21 +9,8 @@
 "use strict";
 
 ((window) => {
+  const core = Deno.core;
   const webidl = window.__bootstrap.webidl;
-  const {
-    forgivingBase64Encode,
-    forgivingBase64Decode,
-  } = window.__bootstrap.infra;
-  const { DOMException } = window.__bootstrap.domException;
-  const {
-    ArrayPrototypeMap,
-    StringPrototypeCharCodeAt,
-    ArrayPrototypeJoin,
-    SafeArrayIterator,
-    StringFromCharCode,
-    TypedArrayFrom,
-    Uint8Array,
-  } = window.__bootstrap.primordials;
 
   /**
    * @param {string} data
@@ -36,13 +23,7 @@
       prefix,
       context: "Argument 1",
     });
-
-    const uint8Array = forgivingBase64Decode(data);
-    const result = ArrayPrototypeMap(
-      [...new SafeArrayIterator(uint8Array)],
-      (byte) => StringFromCharCode(byte),
-    );
-    return ArrayPrototypeJoin(result, "");
+    return core.opSync("op_base64_atob", data);
   }
 
   /**
@@ -56,20 +37,7 @@
       prefix,
       context: "Argument 1",
     });
-    const byteArray = ArrayPrototypeMap(
-      [...new SafeArrayIterator(data)],
-      (char) => {
-        const charCode = StringPrototypeCharCodeAt(char, 0);
-        if (charCode > 0xff) {
-          throw new DOMException(
-            "The string to be encoded contains characters outside of the Latin1 range.",
-            "InvalidCharacterError",
-          );
-        }
-        return charCode;
-      },
-    );
-    return forgivingBase64Encode(TypedArrayFrom(Uint8Array, byteArray));
+    return core.opSync("op_base64_btoa", data);
   }
 
   window.__bootstrap.base64 = {