Follow up to https://github.com/denoland/deno/pull/13839, optimizing `base64_roundtrip` ~20x (~125ms => ~6.5ms)

Before:
```
base64_roundtrip:    	n = 100, dt = 12.493s, r = 8/s, t = 124930000ns/op
```
After:
```
base64_roundtrip:    	n = 100, dt = 0.649s, r = 154/s, t = 6490000ns/op
```